@@ -9,33 +9,35 @@
     "test": "jest"
   },
   "dependencies": {
-    "@emotion/react": "^11.5.0",
-    "@mui/lab": "^5.0.0-alpha.53",
-    "@mui/material": "^5.2.2",
+    "@date-io/date-fns": "^2.14.0",
+    "@emotion/react": "^11.9.3",
+    "@mui/lab": "^5.0.0-alpha.90",
+    "@mui/material": "^5.9.0",
     "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest",
-    "@mui/styles": "^5.0.2",
-    "@mui/x-data-grid": "^5.0.1",
-    "date-fns": "^2.25.0",
-    "history": "5",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-hook-form": "^7.22.2",
-    "react-router-dom": "^6.0.2",
-    "styled-components": "^5.3.3",
-    "use-react-router-breadcrumbs": "^3.0.1"
+    "@mui/styles": "^5.9.0",
+    "@mui/x-data-grid": "^5.13.1",
+    "@mui/x-date-pickers": "^5.0.0-beta.1",
+    "date-fns": "^2.28.0",
+    "history": "5.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.33.1",
+    "react-router-dom": "^6.3.0",
+    "styled-components": "^5.3.5",
+    "use-react-router-breadcrumbs": "^3.2.1"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@vitejs/plugin-react": "^1.0.0",
-    "jest": "^27.3.1",
-    "jest-mock-extended": "^2.0.4",
-    "prettier": "^2.4.1",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.3.2",
-    "vite": "^2.6.4"
+    "@types/jest": "^28.1.6",
+    "@types/node": "^18.0.6",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.0.0",
+    "jest": "^28.1.3",
+    "jest-mock-extended": "^2.0.7",
+    "prettier": "^2.7.1",
+    "ts-jest": "^28.0.7",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.0"
   },
   "resolutions": {
     "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest",
