@@ -1,8 +1,9 @@
 import {Box, Button, Modal, TextField} from '@mui/material'
 import {useState} from 'react'
 import {Controller, SubmitHandler, useForm} from 'react-hook-form'
-import {DateTimePicker, LocalizationProvider} from '@mui/lab'
-import AdapterDateFns from '@mui/lab/AdapterDateFns'
+import {DateTimePicker} from '@mui/x-date-pickers/DateTimePicker'
+import {LocalizationProvider} from '@mui/x-date-pickers'
+import {AdapterDateFns} from '@mui/x-date-pickers/AdapterDateFns'
 
 const style = {
   position: 'absolute' as 'absolute',
@@ -17,7 +18,7 @@ const style = {
 }
 
 type FormInput = {
-  inputValue: Date
+  inputValue: Date | null
 }
 
 const Component = (): JSX.Element => {
@@ -25,7 +26,7 @@ const Component = (): JSX.Element => {
   const handleOpen = () => setOpen(true)
   const handleClose = () => setOpen(false)
 
-  const {control, handleSubmit, setValue} = useForm()
+  const {control, handleSubmit, setValue} = useForm<FormInput>()
   const onSubmit: SubmitHandler<FormInput> = (data) => console.log(data)
 
   return (
